Add files via upload
The first simple implementation enabled the user to move to a consistent input.

It can be moved by w, a, s, and d. It can be moved 20 to 30 times per second.

I set a frame-delay time with pthread, and I have not yet processed the input wait time through the getchar function, so the speed is not as fast as the frame I set.